解决在线迁移任务进度刷新卡在数据校验完成状态的问题

+++ plugins/data-migration/src/main/java/org/opengauss/admin/plugin/domain/MigrationTaskAlertDetail.java
@@ -19,6 +19,6 @@ import lombok.Data;
 @TableName("tb_migration_task_alert_detail")
 public class MigrationTaskAlertDetail {
     @TableId(value = "alert_id")
-    private int alertId;
+    private Long alertId;
     private String detail;
 }
