+++ plugins/data-migration/src/main/java/org/opengauss/admin/plugin/mapper/MigrationTaskAlertMapper.java
@@ -51,5 +51,5 @@ public interface MigrationTaskAlertMapper extends BaseMapper<MigrationTaskAlert>
      * @param alert alert entity
      * @return group alert id list
      */
-    List<Integer> getGroupAlertIds(MigrationTaskAlert alert);
+    List<Long> getGroupAlertIds(MigrationTaskAlert alert);
 }
